Fix NPE on Public Key deletion in case the Environment create failed
and the public key id stays null

+++ environment/src/main/java/com/sequenceiq/environment/environment/flow/deletion/handler/PublicKeyDeleteHandler.java
@@ -68,7 +68,7 @@ public class PublicKeyDeleteHandler extends EventSenderAwareHandler<EnvironmentD
     }
 
     private void deleteManagedKey(Environment environment) {
-        if (environment.getAuthentication().isManagedKey()) {
+        if (environment.getAuthentication().isManagedKey() && environment.getAuthentication().getPublicKeyId() != null) {
             LOGGER.debug("Environment {} has managed public key. Deleting.", environment.getName());
             deletePublicKey(environment);
         } else {
